requirements.txt
python-dotenv~=1.2.1
playwright~=1.58.0
openai~=2.28.0